Fix failing CI test (#20)
The test_cmap_matplotlib_missing test relies on using a color that
branca cannot parse. The new branca release added support for the
color I was using, causing the test to fail. I've switched to a
different color and added a warning for future me if support gets
added for the new color.